webscrapping_join_table_mobile_cpu/requirements.txt
selenium
webdriver-manager
pandas
mysql-connector
gcloud
